requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.1.0
Pillow==10.1.0
django-filter==23.3